feat(auth): allow specifying options without `createAuthStore()`
